Fix accidental overwrite of existing url and push specs
When importing additional repositories on an existing organisation,
the GitHub plugin preserves the legacy 'url' and 'push' specs
for allowing custom ad-hoc settings, such as the use of stable SSH
keys for replication or more restrictive push specs.

The abstraction of the replication configuration update implemented
with I9bde7e1148 has accidental broken the current behaviour causing
a regression in how the remotes are configured.

Reintroduce the correct behaviour by reading the existing config
when updating a remote and avoid overwriting the url and push
replication settings.

Change-Id: Idd7113f6c0549d752b1dd31bf52d960aee1f77b1