Use Java record-style methods for accessing Scope
The introduction of the Java 17 record type for Scope
with I979fbfe27 has caused incompatiblity with the current
Velocity templates because the fields are now accessible
as methods.

Adapt the Velocity template calls in scope.html to resume
the current scopes selection functionality.

Also create the missing getter for the Scope enum in
OAuthProtocol for allowing Velocity to access it.

Change-Id: I8f6e404c4a3f401deafa25061c1749028319e5b4